refactor: de-part diplomacy_panel cluster into explicit-import libraries (#4117)
Convert the 12-file diplomacy_panel part cluster to standalone libraries
with widget/state split and shared chrome-colors module. Shrink
repo.app_no_part_directives grandfather by twelve entries; behavior unchanged.

Refs #4117
