+++ app/lib/features/game/widgets/diplomacy/diplomacy_panel_chrome_section_header.dart
@@ -2,24 +2,13 @@
 // SPEC/ui/diplomacy-panel.md § Section headings and § Per-faction row → Type
 // badge colors.
 
-part of 'diplomacy_panel.dart';
+import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+import 'package:flutter/material.dart';
 
-/// Translucent overlay used as the WAR-state badge background, derived
-/// from the `--danger` hue at lightness 0.40, chroma 0.06, hue 20, alpha
-/// 0.40. SPEC/ui/diplomacy-panel.md § Relation state badge cites the
-/// mockup token `oklch(40% 0.06 20 / 0.4)` directly; this constant
-/// computes the same OKLCH sRGB approximation through [oklchToColor].
-final Color _kWarBadgeBackground = oklchToColor(
-  const OklchToken(0.40, 0.06, 20),
-).withValues(alpha: 0.4);
-
-/// Translucent overlay used as the PEACE-state badge background, derived
-/// from the `--success` hue at lightness 0.40, chroma 0.06, hue 150,
-/// alpha 0.20. SPEC/ui/diplomacy-panel.md § Relation state badge cites
-/// the mockup token `oklch(40% 0.06 150 / 0.2)`.
-final Color _kPeaceBadgeBackground = oklchToColor(
-  const OklchToken(0.40, 0.06, 150),
-).withValues(alpha: 0.2);
+import '../../../../config/themes.dart' show editorialMonocleDisplayFontFamily;
+import '../../../../widgets/ct_spacing.dart';
+import 'diplomacy_panel_rows.dart';
 
 /// Section heading for a diplomacy faction group (Great Powers / Minor
 /// Nations / Tribes).
@@ -28,8 +17,8 @@ final Color _kPeaceBadgeBackground = oklchToColor(
 /// text color, 2 px `--accent-dim` bottom border per
 /// [mockups/GAME30001-diplomacy-panel.html](../../../../../SPEC/ui/mockups/GAME30001-diplomacy-panel.html)
 /// `.section-head`.
-class _DiplomacySectionHeader extends StatelessWidget {
-  const _DiplomacySectionHeader({required this.title, this.isFirst = false});
+class DiplomacySectionHeader extends StatelessWidget {
+  const DiplomacySectionHeader({required this.title, this.isFirst = false});
 
   final String title;
 
@@ -85,8 +74,8 @@ class _DiplomacySectionHeader extends StatelessWidget {
 /// `.f-badge` chrome (mono font, tight letter-spacing, square `1px`
 /// border-radius). All colors resolve from the canonical editorial-monocle
 /// palette — no hardcoded Material chrome.
-class _FactionKindBadge extends StatelessWidget {
-  const _FactionKindBadge({required this.kind});
+class DiplomacyFactionKindBadge extends StatelessWidget {
+  const DiplomacyFactionKindBadge({required this.kind});
 
   final FactionKind kind;
